Compiled raw RV ASM through cargo to blink
Sucessfully compiled raw RISC-V assembly through cargo,
then transferred to MQ Pro, to blink. The ASM and
linker script was stolen from CS 107E, and various sources
were used to enable the correct toolchains and options.

In process, switched to nightly compiler, but seems like
in the end was actually unnecessary. The problem ended
up being that the sections were compiled directly to what
the linker set them to, so it seems like the entire
`0x50000000` space was being taken up. Not knowing this,
I set out to try and desparately reduce the binary. After
cutting down the linker to just `.text`, ended up working.

TODO: Figure out how to reinstate the required memory
placement without bloating the binary, because the transfer
speed of data to MQ Pro (in this current setup) is too
atrocious to copy the entire image. This will definitely
require rechecking CS 107E.